change opal_value_t array structs to use int32_t for size for compatibility with buffer packing, add pack/unpack routines for new opal_value_t array types

+++ opal/dss/dss_types.h
@@ -118,82 +118,82 @@ typedef uint64_t opal_identifier_t;
 /* List types for opal_value_t, needs number of elements and a pointer */
 /* float array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     float *data;
 } opal_float_array_t;
 /* double array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     double *data;
 } opal_double_array_t;
 /* string array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     char **data;
 } opal_string_array_t;
 /* bool array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     bool *data;
 } opal_bool_array_t;
 /* size array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     size_t *data;
 } opal_size_array_t;
 /* opal byte object array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     opal_byte_object_t *data;
 } opal_byte_object_array_t;
 /* int array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     int *data;
 } opal_int_array_t;
 /* int8 array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     int8_t *data;
 } opal_int8_array_t;
 /* int16 array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     int16_t *data;
 } opal_int16_array_t;
 /* int32 array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     int32_t *data;
 } opal_int32_array_t;
 /* int64 array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     int64_t *data;
 } opal_int64_array_t;
 /* uint array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     unsigned int *data;
 } opal_uint_array_t;
 /* uint8 array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     uint8_t *data;
 } opal_uint8_array_t;
 /* uint16 array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     uint16_t *data;
 } opal_uint16_array_t;
 /* uint32 array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     uint32_t *data;
 } opal_uint32_array_t;
 /* uint64 array object */
 typedef struct {
-    size_t size;
+    int32_t size;
     uint64_t *data;
 } opal_uint64_array_t;
